+++ extensions/ql-vscode/src/language-support/contextual/key-type.ts
