+++ extensions/ql-vscode/test/vscode-tests/no-workspace/language-support/contextual/ast-builder.test.ts
@@ -1,10 +1,10 @@
 import { readFileSync } from "fs-extra";
 
-import AstBuilder from "../../../../src/contextual/astBuilder";
-import { CodeQLCliServer } from "../../../../src/cli";
+import AstBuilder from "../../../../../src/language-support/contextual/ast-builder";
+import { CodeQLCliServer } from "../../../../../src/cli";
 import { Uri } from "vscode";
-import { QueryOutputDir } from "../../../../src/run-queries-shared";
-import { mockDatabaseItem, mockedObject } from "../../utils/mocking.helpers";
+import { QueryOutputDir } from "../../../../../src/run-queries-shared";
+import { mockDatabaseItem, mockedObject } from "../../../utils/mocking.helpers";
 import path from "path";
 
 /**
