+++ extensions/ql-vscode/test/vscode-tests/no-workspace/run-queries.test.ts
@@ -11,7 +11,7 @@ import {
 import * as config from "../../../src/config";
 import { tmpDir } from "../../../src/helpers";
 import { CodeQLCliServer } from "../../../src/cli";
-import { SELECT_QUERY_NAME } from "../../../src/contextual/locationFinder";
+import { SELECT_QUERY_NAME } from "../../../src/language-support/contextual/location-finder";
 import {
   QueryInProgress,
   compileQuery as compileQueryLegacy,
